Move service host to modes/shared
